+++ src/imagemin-loader.js
@@ -6,18 +6,19 @@ const minify = require("./minify/minify");
 
 module.exports = function(content) {
   const options = loaderUtils.getOptions(this) || {};
-  const bail =
-    options.bail ||
-    /* eslint-disable-next-line no-underscore-dangle */
-    (this._compiler && this._compiler.options && this._compiler.options.bail) ||
-    false;
-  const plugins = options.plugins || [];
   const callback = this.async();
 
-  return nodeify(
-    minify(content, { bail, imageminOptions: { plugins } }),
-    callback
-  );
+  if (typeof options.bail === "undefined") {
+    /* eslint-disable no-underscore-dangle */
+    options.bail =
+      (this._compiler &&
+        this._compiler.options &&
+        this._compiler.options.bail) ||
+      false;
+    /* eslint-enable no-underscore-dangle */
+  }
+
+  return nodeify(minify(content, options), callback);
 };
 
 module.exports.raw = true;